requirements/base.txt
Django==4.1.3
drf_yasg==1.21.6
python-dotenv==1.0.0
Requests==2.31.0
